requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-pinecone==0.2.8

langchain-google-genai>=2.0.0

google-generativeai>=0.7.0,<0.8.0

flask==3.1.1
python-dotenv==1.1.0
gunicorn
-e .


# older version
# langchain==0.3.26
# flask==3.1.1
# sentence-transformers==4.1.0
# pypdf==5.6.1
# python-dotenv==1.1.0
# langchain-pinecone==0.2.8
# # langchain-openai==0.3.24
# langchain-community==0.3.26
# # Gemini integration
# # langchain-google-genai==1.0.6
# google-generativeai==0.5.4
# gunicorn
# langchain-google-genai


# -e .